- Shared Admin and Staff Layout - Administration nav bar responsiveness - Registration Form mobile UI
