+++ RippedAndFit.Web/Views/Admin/AddStaff.cshtml
@@ -1,6 +1,5 @@
-﻿    @{
+﻿@{
     ViewData["Title"] = "Add Staff";
-    Layout = "~/Views/Shared/_AdminLayout.cshtml";
 }
 
 <h1>
