--- RippedAndFit.Web/Views/Admin/Logs.cshtml
@@ -1,8 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Logs";
-    Layout = "~/Views/Shared/_AdminLayout.cshtml";
-}
-
-<h1>
-    Logs
-</h1>
