--- RippedAndFit.Web/Views/Admin/Registration.cshtml
@@ -1,64 +1,0 @@
-﻿@using RippedAndFit.Domain.Enums
-
-@model RippedAndFit.Domain.Entities.NewMember;
-
-@{
-    ViewData["Title"] = "Registration";
-    Layout = "~/Views/Shared/_AdminLayout.cshtml"; 
-}
-
-<h1>Register Member</h1>
-
-<div>
-    <form method="post">
-        <div class="row-1">
-            <label>Name</label>
-            <input asp-for="MemberDetails.FirstName" placeholder="First Name" />
-            <input asp-for="MemberDetails.LastName" placeholder="Last Name" />
-        </div>
-        <div class="row-2">
-            <label>Sex</label>
-            <div>
-                <input asp-for="MemberDetails.Gender" value="Male" type="radio" name="gender" id="male"/>
-                <label for="male">Male</label>
-                <input asp-for="MemberDetails.Gender" value="Female" type="radio" name="gender" id="female"/>
-                <label for="female">Female</label>
-            </div>
-            <label>Date of Birth</label>
-            <input asp-for="MemberDetails.DateOfBirth" placeholder="DD-MM-YYYY" type="date" />
-        </div>
-        <div class="row-3">
-            <label>Phone</label>
-            <input asp-for="MemberDetails.PhoneNumber" placeholder="" />
-            <label>Age</label>
-            <input asp-for="MemberDetails.Age" placeholder="" />
-        </div>
-        <div>
-            <label>Email</label>
-            <input asp-for="MemberDetails.Email"/>
-            <span asp-validation-for="MemberDetails.Email"></span>
-        </div>
-        <div class="row-5">
-            <label>Member Type</label>
-            <select asp-for="MemberDetails.MemberType">
-                @foreach (var memberType in Enum.GetValues(typeof(MemberType)))
-                {
-                    <option value="@memberType">@memberType</option>
-                }
-            </select>
-        </div>
-        
-        <div class="row-1">
-            <label>Username</label>
-            <input asp-for="User.Username" />
-            <span asp-validation-for="User.Username"></span>
-        </div>
-        <div class="row-1">
-            <label>Password</label>
-            <input asp-for="User.Password" type="password" />
-        </div>
-        <div>
-            <button type="submit">Register</button>
-        </div>
-    </form>
-</div>
