+++ RippedAndFit.Web/Views/Admin/Staffs.cshtml
@@ -3,7 +3,6 @@
 
 @{
     ViewData["Title"] = "Staffs";
-    Layout = "~/Views/Shared/_AdminLayout.cshtml";
 }
 
 @model RippedAndFit.Web.Models.StaffsModel;
