+++ RippedAndFit.Web/Views/Admin/UpdateStaff.cshtml
@@ -1,6 +1,5 @@
 ﻿@{
     ViewData["Title"] = "Update Staff";
-    Layout = "~/Views/Shared/_AdminLayout.cshtml";
 }
 
 <h1>
