+++ RippedAndFit.Web/Views/Administration/Dashboard.cshtml
@@ -1,6 +1,5 @@
 ﻿@{
     ViewData["Title"] = "Dashboard";
-    Layout = "~/Views/Shared/_AdminLayout.cshtml";
 }
 
 <h1>
