+++ RippedAndFit.Web/Views/Administration/Logs.cshtml
@@ -1,6 +1,5 @@
 ﻿@{
     ViewData["Title"] = "Logs";
-    Layout = "~/Views/Shared/_StaffLayout.cshtml";
 }
 
 <h1>
