+++ RippedAndFit.Web/Views/Administration/Profile.cshtml
@@ -1,0 +1,7 @@
+﻿@{
+    ViewData["Title"] = "Profile";
+}
+
+<h1>
+    Profile
+</h1>
