+++ RippedAndFit.Web/Views/Administration/UpdateMember.cshtml
@@ -1,6 +1,5 @@
 ﻿@{
     ViewData["Title"] = "Update Member";
-    Layout = "~/Views/Shared/_AdminLayout.cshtml";
 }
 
 @model RippedAndFit.Domain.Entities.Users;
