+++ RippedAndFit.Web/Views/Home/About.cshtml
@@ -1,5 +1,6 @@
 ﻿@{
     ViewData["Title"] = "About Us";
+    Layout = "~/Views/Shared/_HomeLayout.cshtml";
 }
 
 <div>
