+++ RippedAndFit.Web/Views/Home/Contact.cshtml
@@ -1,5 +1,6 @@
 ﻿@{
     ViewData["Title"] = "Contact Us";
+    Layout = "~/Views/Shared/_HomeLayout.cshtml";
 }
 
 <div>
