+++ RippedAndFit.Web/Views/Home/Home.cshtml
@@ -1,5 +1,6 @@
 ﻿@{
     ViewData["Title"] = "Home";
+    Layout = "~/Views/Shared/_HomeLayout.cshtml";
 }
 
 <div class="home-background">
