+++ RippedAndFit.Web/Views/Home/Login.cshtml
@@ -2,6 +2,7 @@
 
 @{
     ViewData["Title"] = "Login";
+    Layout = "~/Views/Shared/_HomeLayout.cshtml";
 }
 
 <div class="home-background">
@@ -12,6 +13,8 @@
         </h1>
         <img src="~/image/Push-Up.svg" alt="Guy doing a Push-Up"/>
     </div>
+
+    <div class="login-form-bg"></div>
 
     <div class="login-form">
         <form method="post">
@@ -35,5 +38,4 @@
             </div>
         </form>
     </div>
-
 </div>
