+++ RippedAndFit.Web/Views/Member/Dashboard.cshtml
@@ -1,6 +1,5 @@
 ﻿@{
     ViewData["Title"] = "Dashboard";
-    Layout = "~/Views/Shared/_MemberLayout.cshtml";
 }
 
 <h1>
