+++ RippedAndFit.Web/Views/Shared/_AdministrationLayout.cshtml
@@ -8,7 +8,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
-    <link rel="stylesheet" href="~/css/administration-navbar.css" />
+    <link rel="stylesheet" href="~/css/administration-navbar.css"/>
     <link rel="stylesheet" href="~/css/registration.css" />
 </head>
 <body>
@@ -24,7 +24,7 @@
 
         <div class="nav-bar">
             <div>
-                <a class="menu-close hide-desktop">
+                <a class="menu-close">
                     <img src="@Url.Content("~/image/Close-Icon.svg")" alt="Close Menu" />
                 </a>
             </div>
@@ -34,7 +34,6 @@
             <div class="nav-tab">
                 <a asp-controller="Administration" asp-action="Dashboard"><i class="bi bi-window-desktop"></i> Dashboard</a>
                 <a asp-controller="Administration" asp-action="Registration"><i class="bi bi-person-lines-fill"></i> Registration</a>
-                <a asp-controller="Admin" asp-action="Staffs"><i class="bi bi-person-fill-gear"></i> Staffs</a>
                 <a asp-controller="Administration" asp-action="Members"><i class="bi bi-people-fill"></i> Members</a>
                 <a asp-controller="Administration" asp-action="Logs"><i class="bi bi-clock-history"></i> Logs</a>
             </div>
@@ -44,7 +43,7 @@
             </div>
         </div>
 
-        <div class="main-content">
+        <div class="home-content">
             <main role="main">
                 @RenderBody()
             </main>
