+++ RippedAndFit.Web/Views/Shared/_HomeLayout.cshtml
@@ -22,13 +22,13 @@
             <nav>
                 <div>
                     <a href="#" class="menu-open hide-desktop">
-                        <img src="@Url.Content("~/image/Menu-Icon.png")" alt="Open Menu" />
+                        <img src="@Url.Content("~/image/Menu-Icon.svg")" alt="Open Menu" />
                     </a>
                 </div>
                 <ul>
                     <li class="hide-desktop">
                         <a href="#" class="menu-close">
-                            <img src="@Url.Content("~/image/Close-Icon.png")" alt="Close Menu" />
+                            <img src="@Url.Content("~/image/Close-Icon.svg")" alt="Close Menu" />
                         </a>
                     </li>
                     <li><a asp-controller="Home" asp-action="Home">Home</a></li>
