--- RippedAndFit.Web/Views/Shared/_StaffLayout.cshtml
@@ -1,39 +1,0 @@
-﻿<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>@ViewData["Title"] - Ripped And Fit</title>
-</head>
-<body>
-    <header>
-        <h1>RIPPED AND FIT GYM</h1>
-    </header>
-    <div>
-        <nav>
-            <div>
-                <ul>
-                    <li>
-                        <a asp-controller="Staff" asp-action="Dashboard">Dashboard</a>
-                    </li>
-                    <li>
-                        <a asp-controller="Staff" asp-action="Registration">Registration</a>
-                    </li>
-                    <li>
-                        <a asp-controller="Staff" asp-action="Members">Members</a>
-                    </li>
-                    <li>
-                        <a asp-controller="Staff" asp-action="Logs">Logs</a>
-                    </li>
-                </ul>
-            </div>
-        </nav>
-    </div>
-    <div>
-        <main role="main">
-            @RenderBody()
-        </main>
-    </div>
-    @await RenderSectionAsync("Scripts", required: false)
-</body>
-</html>
