--- RippedAndFit.Web/Views/Staff/Dashboard.cshtml
@@ -1,8 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Dashboard";
-    Layout = "~/Views/Shared/_StaffLayout.cshtml";
-}
-
-<h1>
-    Hi, Staff
-</h1>
