--- RippedAndFit.Web/Views/Staff/Members.cshtml
@@ -1,8 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Members";
-    Layout = "~/Views/Shared/_StaffLayout.cshtml";
-}
-
-<h1>
-    Members
-</h1>
