+++ RippedAndFit.Web/wwwroot/css/administration-navbar.css
@@ -1,0 +1,119 @@
+﻿.hide-mobile {
+    display: none;
+}
+
+.gym-icon {
+    background-color: #29292e;
+    text-align: center;
+}
+
+.gym-icon img {
+    width: auto;
+    height: auto;
+    max-width: 175px;
+}
+
+header {
+    background: #202020;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    height: 60px;
+}
+
+body {
+    margin: 0;
+    font-family: 'Poppins';
+    overflow: hidden;
+}
+
+.nav-bar {
+    background: #202020;
+    position: fixed;
+    z-index: 999;
+    top: 0;
+    left: 0;
+    height: 100vh;
+    margin: 0;
+    width: 300px;
+    transform: translateX(-100%);
+    transition: transform 0.5s ease-in-out;
+}
+
+.nav-tab {
+    padding: 0.5em;
+    display: flex;
+    flex-direction: column;
+    margin: 1.2em auto;
+}
+
+.nav-tab a {
+    text-decoration: none;
+    color: white;
+    font-size: 1.7rem;
+    padding: 0.5em;
+    transition: transform 0.5s ease-in-out;
+}
+
+.nav-tab a:hover {
+    font-size: 2rem;
+}
+
+.nav-tab a i {
+    color: #ed563b;
+}
+
+.utility {
+    padding: 0.5em;
+    position: fixed;
+    bottom: 0;
+    display: flex;
+    flex-direction: column;
+    margin: 1.2em auto;
+}
+
+.utility a {
+    text-decoration: none;
+    color: white;
+    font-size: 1.7rem;
+    padding: 0.5em;
+    transition: transform 0.5s ease-in-out;
+}
+
+.utility a:hover {
+    font-size: 2rem;
+}
+
+.utility a i {
+    color: #ed563b;
+}
+
+.nav-bar.open {
+    transform: translateX(0);
+}
+
+@media (min-width: 1080px) {
+    .hide-mobile {
+        display: block;
+    }
+
+    .hide-desktop {
+        display: none;
+    }
+
+    .nav-bar {
+        transform: translateX(0);
+        position: fixed;
+        bottom: 0;
+        margin-top: 60px;
+        height: calc(100vh - 60px);
+    }
+
+    .main-content {
+        position: fixed;
+        margin-left: 300px;
+        width: calc(100vw - 300px);
+    }
+}
+
+
