+++ RippedAndFit.Web/wwwroot/css/header.css
@@ -16,7 +16,7 @@ header {
     display: flex;
     align-items: center;
     justify-content: space-between;
-    background: #343434;
+    background: #202020;
     height: 60px;
 }
 
@@ -32,7 +32,7 @@ header {
 }
 
 nav ul {
-    background: #343434;
+    background: #202020;
     position: fixed;
     z-index: 999;
     top: 0;
